Provide tree depth on the LoaderContext while loading sourcemaps
It's important to know both the importer and the depth of the sourcemap tree when loading, to
solve cases like https://github.com/babel/babel/pull/14274. Here, a parent sourcemap contains a
source location that, when resolved, has the same location as the parent. This easily happens when
the file represents an in-place transformation of the original sourcemap (eg, `babel script.js
--out-file script.js --source-maps`). Because they exist in the same location, it's extremely easy
to create a loader which continues to return the transformed sourcemap when we intend to only return
a `null` (no sourcemap) for the original source file.