fix: pathological slow case in bySource
Hacked together a bit to fix #41. The use of a proxy here is fairly
cursed but I wanted to validate if this approach would fix it, which
it does -- tests pass and the 24s case goes is 90ms on my machine.

Feel free to rewrite or suggest a cleaner approach :stuck_out_tongue: